Replace height with max-height
Currently, we are using "max-width: 90%;" as part of .post-content img
that forces a maximum width. Together with a hardcoded height, this can
lead to distorted images.
When setting max-height instead of height on an image, either of the two
max settings will take precedence over the other and scale the image while
keeping the aspect ratio fixed.